requirements.txt
# ============================================
# KG Gap Discovery - Python Dependencies
# ============================================
# Install: pip install -r requirements.txt
# Python version: 3.9+ recommended
# ============================================

# --- Data Collection ---
requests>=2.31.0
tqdm>=4.66.0

# --- Text Processing ---
PyMuPDF>=1.23.0            # PDF text extraction (import as fitz)
spacy>=3.7.0               # NLP preprocessing

# --- LLM API ---
openai>=1.12.0             # For GPT-4 triple extraction
# anthropic>=0.18.0        # Uncomment if using Claude instead

# --- Knowledge Graph ---
networkx>=3.2.0            # Graph construction and analysis
python-louvain>=0.16       # Community detection (import as community)

# --- Entity Deduplication ---
sentence-transformers>=2.3.0   # Sentence-BERT embeddings
fuzzywuzzy>=0.18.0             # Fuzzy string matching
python-Levenshtein>=0.23.0     # Speed up fuzzywuzzy

# --- Graph Embeddings ---
pykeen>=1.10.0             # TransE and other KG embedding models
torch>=2.1.0               # PyKEEN dependency (CPU version is fine)

# --- Visualisation ---
pyvis>=0.3.2               # Interactive graph visualisation
matplotlib>=3.8.0          # Static plots and charts
seaborn>=0.13.0            # Statistical visualisation

# --- Data Handling ---
pandas>=2.1.0              # DataFrames for tabular data
numpy>=1.26.0              # Numerical operations
scikit-learn>=1.3.0        # Metrics, cosine similarity

# --- Config ---
pyyaml>=6.0.1              # YAML config parsing

# --- Notebooks ---
jupyter>=1.0.0             # Jupyter notebook support
ipywidgets>=8.1.0          # Interactive widgets in notebooks
